+ Switched to double precision + Enabled OpenMP + Added some progress feedback

+++ boundingBox.cpp
@@ -1,12 +1,12 @@
 #include "boundingBox.h"
-BoundingBox::BoundingBox(float *p1, float *p2)
+BoundingBox::BoundingBox(double *p1, double *p2)
 {
   // Enforce lower left corner is p1 and upper right is p2
   for(int i=0; i<3; i++)
   {
     if(p1[i]>p2[i])
     {
-      float tmp = p1[i];
+      double tmp = p1[i];
       p1[i] = p2[i];
       p2[i] = tmp;
     }
