+++ shapes/cylinder.h
@@ -8,11 +8,11 @@ class Cylinder: public Shape {
      * Cylinder constructor
      * @param state State to be applied to particles within this shape. state = 0 will remove particle from output
      * @param c Center of the cylinder in cartesian coordinates
-     * @param r Radius of the cylinder 
-     * @param l Length of the cylinder in cartesian coordinates, (0,0,l) or (0,l,0) or (l,0,0) indicates the direction 
+     * @param r Radius of the cylinder
+     * @param l Length of the cylinder in cartesian coordinates, (0,0,l) or (0,l,0) or (l,0,0) indicates the direction
      * @param domain Domain in which this shape will be mapped
      */
-    Cylinder(int state, float *c, float r, float *l);
+    Cylinder(int state, double *c, double r, double *l);
     ~Cylinder();
 
     /**
@@ -20,12 +20,12 @@ class Cylinder: public Shape {
      * @param  pt Coordinates to check
      * @return    True if coordinate is inside shape, false otherwise
      */
-    bool IsInside(float *pt);
+    bool IsInside(double *pt);
 
   private:
-    float *c; /**< Center of the cylinder in cartesian coordinates */
-    float r; /**< Radius of the cylinder */
-    float *l; /**< Length of the cylinder */
+    double *c; /**< Center of the cylinder in cartesian coordinates */
+    double r; /**< Radius of the cylinder */
+    double *l; /**< Length of the cylinder */
 };
 
 #endif
