+++ shapes/sphere.h
@@ -8,10 +8,10 @@ class Sphere: public Shape {
      * Sphere constructor
      * @param state State to be applied to particles within this shape. state = 0 will remove particle from output
      * @param c Center of the sphere in cartesian coordinates
-     * @param r Radius of the sphere 
+     * @param r Radius of the sphere
      * @param domain Domain in which this shape will be mapped
      */
-    Sphere(int state, float *c, float r);
+    Sphere(int state, double *c, double r);
     ~Sphere();
 
     /**
@@ -19,11 +19,11 @@ class Sphere: public Shape {
      * @param  pt Coordinates to check
      * @return    True if coordinate is inside shape, false otherwise
      */
-    bool IsInside(float *pt);
+    bool IsInside(double *pt);
 
   private:
-    float *c; /**< Center of the sphere in cartesian coordinates */
-    float r; /**< Radius of the sphere */
+    double *c; /**< Center of the sphere in cartesian coordinates */
+    double r; /**< Radius of the sphere */
 };
 
 #endif
